refactor(control): split plan-todo seeding and module detection out of controller.go
Pure code motion: the plan-todo seeding cluster (seedPlanTodos,
PlanTodosJSON, parsePlanTodos, listItem and friends) moves to
plan_todos.go and the /plan-exec workspace module scan
(detectProjectModules and helpers) moves to project_modules.go.
controller.go 6818 -> 6498 lines, no signature or behavior changes.
